fixed plugins

+++ plugins/plugins-available/mobile/lib/Thruk/Controller/mobile.pm
@@ -30,6 +30,7 @@ page: /thruk/cgi-bin/mobile.cgi
 =cut
 sub mobile_cgi : Regex('thruk\/cgi\-bin\/mobile\.cgi') {
     my ( $self, $c ) = @_;
+    return if defined $c->{'cancled'};
     return $c->detach('/mobile/index');
 }
