+++ plugins/plugins-available/statusmap/lib/Thruk/Controller/statusmap.pm
@@ -31,6 +31,7 @@ page: /thruk/cgi-bin/statusmap.cgi
 =cut
 sub statusmap_cgi : Regex('thruk\/cgi\-bin\/statusmap\.cgi') {
     my ( $self, $c ) = @_;
+    return if defined $c->{'cancled'};
     return $c->detach('/statusmap/index');
 }
